Move generalized read function to io.__init__. This fixes an import problem in python 3.6.
